feat(Spinner.tsx): add Spinner component to display a loading indicator in the app
The Spinner component is a new component that displays a loading indicator using the ActivityIndicator component from the react-native library. It is used to indicate that the app is currently loading or performing a task. The Spinner component is styled using the StyleSheet.create function to set the margin vertical spacing to 20.