feat(theme): add success and successGreyedOut colors to customLightThemeColors and customDarkThemeColors
The success color is added to represent a successful state or action in the UI. The successGreyedOut color is added to represent a disabled or inactive state of a successful action. These colors will be used in various components and elements throughout the application to provide visual feedback to the user.