[FEATURE] Loading spinner and no habit label (#33)
* feat(StoreContext.tsx): add 'loading' property to the initial state in StoreContext to track loading state
feat(reducer.ts): add 'SET_LOADING' action to update the 'loading' state in the reducer

* feat(Spinner.tsx): add Spinner component to display a loading indicator in the app

The Spinner component is a new component that displays a loading indicator using the ActivityIndicator component from the react-native library. It is used to indicate that the app is currently loading or performing a task. The Spinner component is styled using the StyleSheet.create function to set the margin vertical spacing to 20.

* feat(AppTabs.tsx): add error handling and loading state management when fetching habits and user from the server

The changes in this commit add error handling and loading state management when fetching habits and user from the server in the `AppTabs.tsx` file. Previously, the habits and user were fetched without any error handling or loading state management.

Now, when fetching habits and user, the code is wrapped in a try-catch block. If an error occurs during the fetch, it will be caught and logged to the console. Additionally, a loading state is set to true before the fetch and set to false after the fetch, ensuring that the loading state is properly managed.

These changes improve the robustness and user experience of the application by handling errors gracefully and providing feedback to the user during the fetch process.

* feat(components): add Spinner component to indicate loading state in HabitsScreen, SettingsScreen, and TodayScreen

The Spinner component is added to the HabitsScreen, SettingsScreen, and TodayScreen components to indicate the loading state. This will provide visual feedback to the user while data is being fetched or processed.

* fix(HabitsScreen.tsx): add conditional rendering for displaying a message when there are no habits
fix(TodayScreen.tsx): add conditional rendering for displaying a message when there are no habits for today
fix(index.ts): add new labels for displaying messages when there are no habits
fix(axios.ts): temporarily return an empty array instead of mockHabits for /api/habits endpoint

* fix: mAJOR bug: stuck on screens requiring params after logging out and in again

* refactor(HabitsStack.tsx): remove unused imports and unnecessary options in createStackNavigator
refactor(SettingsStack.tsx): remove unused imports and unnecessary options in createStackNavigator