fix(PressableCard.tsx): import useCustomTheme from @app/theme/useCustomTheme instead of customLightThemeColors to improve code organization and maintainability feat(PressableCard.tsx): update PressableCard component to change background color and opacity when pressed or isGreyedOut is true, using the theme colors fix(HabitsScreen.tsx): import useCustomTheme from @app/theme/useCustomTheme instead of useTheme from react-native-paper to improve code organization and maintainability fix(SettingsScreen.tsx): import useCustomTheme from @app/theme/useCustomTheme instead of customLightThemeColors to improve code organization and maintainability fix(SettingsScreen.tsx): remove unused styles and move styles declaration inside the component function to improve code organization fix(TodayHabit.tsx): import useCustomTheme from @app/theme/useCustomTheme instead of customLightThemeColors to improve code organization and maintainability feat(TodayHabit.tsx): update TodayHabit component to change habit name and daysDue color based on isGreyedOut using the theme colors
fix(theme): update primary and success colors in customDarkThemeColors

The primary color in the customDarkThemeColors object was updated from 'rgb(183,196,255)' to 'rgb(183,208,255)' to improve the visual appearance.

Additionally, the success color and successGreyedOut color were swapped. The success color was updated from 'rgb(121, 184, 149)' to 'rgb(45, 181, 106)', and the successGreyedOut color was updated from 'rgb(45, 181, 106)' to 'rgb(121, 184, 149)' to ensure consistency and clarity in the theme colors.

The grey color shades in the customDarkThemeColors object were also adjusted for better visual contrast. The values for shades 100, 200, 300, and 400 were updated to 'rgb(56,59,65)', 'rgb(85,87,98)', 'rgb(104,108,117)', and 'rgb(141,148,155)' respectively.